Added TryGetValue directly to PropertyValues
