--- build.cmd
@@ -1,3 +1,0 @@
-@echo off
-powershell -ExecutionPolicy ByPass -NoProfile -command "& '%~dp0eng\common\Build.ps1' -restore -build %*"
-exit /b %ErrorLevel%
