--- src/EFCore.Sqlite/EFCore.Sqlite.csproj
@@ -1,39 +1,0 @@
-﻿<Project>
-
-  <Import Project="Sdk.props" Sdk="Microsoft.NET.Sdk" />
-
-  <PropertyGroup>
-    <Description>SQLite database provider for Entity Framework Core.</Description>
-    <TargetFramework>netstandard2.1</TargetFramework>
-    <MinClientVersion>3.6</MinClientVersion>
-    <PackageId>Microsoft.EntityFrameworkCore.Sqlite</PackageId>
-    <PackageTags>$(PackageTags);SQLite</PackageTags>
-    <IncludeBuildOutput>false</IncludeBuildOutput>
-    <IncludeSymbols>false</IncludeSymbols>
-    <DisableImplicitFrameworkReferences>true</DisableImplicitFrameworkReferences>
-    <CodeAnalysisRuleSet>..\..\EFCore.ruleset</CodeAnalysisRuleSet>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\EFCore.Sqlite.Core\EFCore.Sqlite.Core.csproj" PrivateAssets="contentfiles;build" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="SQLitePCLRaw.bundle_e_sqlite3" Version="$(SQLitePCLRawBundleESqlite3PackageVersion)">
-      <PrivateAssets>None</PrivateAssets>
-    </PackageReference>
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="lib\**\*">
-      <Pack>True</Pack>
-      <PackagePath>lib</PackagePath>
-    </None>
-  </ItemGroup>
-
-  <Import Project="Sdk.targets" Sdk="Microsoft.NET.Sdk" />
-
-  <Target Name="Compile" />
-  <Target Name="CopyFilesToOutputDirectory" />
-
-</Project>
