--- src/EFCore/ChangeTracking/Internal/InternalShadowEntityEntry.cs
@@ -1,160 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System.Collections.Generic;
-using JetBrains.Annotations;
-using Microsoft.EntityFrameworkCore.Internal;
-using Microsoft.EntityFrameworkCore.Metadata;
-using Microsoft.EntityFrameworkCore.Metadata.Internal;
-using Microsoft.EntityFrameworkCore.Storage;
-
-namespace Microsoft.EntityFrameworkCore.ChangeTracking.Internal
-{
-    /// <summary>
-    ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-    ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-    ///     any release. You should only use it directly in your code with extreme caution and knowing that
-    ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-    /// </summary>
-    public class InternalShadowEntityEntry : InternalEntityEntry
-    {
-        private readonly ISnapshot _propertyValues;
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        public override object Entity => null;
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        public InternalShadowEntityEntry(
-            [NotNull] IStateManager stateManager,
-            [NotNull] IEntityType entityType)
-            : base(stateManager, entityType)
-        {
-            _propertyValues = entityType.GetEmptyShadowValuesFactory()();
-
-            // ReSharper disable once DoNotCallOverridableMethodsInConstructor
-            MarkShadowPropertiesNotSet(entityType);
-        }
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        public InternalShadowEntityEntry(
-            [NotNull] IStateManager stateManager,
-            [NotNull] IEntityType entityType,
-            in ValueBuffer valueBuffer)
-            : base(stateManager, entityType)
-        {
-            _propertyValues = ((EntityType)entityType).ShadowValuesFactory(valueBuffer);
-        }
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        protected override T ReadShadowValue<T>(int shadowIndex)
-            => _propertyValues.GetValue<T>(shadowIndex);
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        protected override object ReadPropertyValue(IPropertyBase propertyBase)
-            => _propertyValues[propertyBase.GetShadowIndex()];
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        protected override bool PropertyHasDefaultValue(IPropertyBase propertyBase)
-            => propertyBase.ClrType.IsDefaultValue(_propertyValues[propertyBase.GetShadowIndex()]);
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        protected override void WritePropertyValue(IPropertyBase propertyBase, object value, bool forMaterialization)
-            => _propertyValues[propertyBase.GetShadowIndex()] = value;
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        public override object GetOrCreateCollection(INavigation navigation, bool forMaterialization)
-            => GetOrCreateCollectionTyped(navigation);
-
-        private ICollection<object> GetOrCreateCollectionTyped(INavigation navigation)
-        {
-            if (!(_propertyValues[navigation.GetShadowIndex()] is ICollection<object> collection))
-            {
-                collection = new HashSet<object>();
-                _propertyValues[navigation.GetShadowIndex()] = collection;
-            }
-
-            return collection;
-        }
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        public override bool CollectionContains(INavigation navigation, InternalEntityEntry value)
-            => GetOrCreateCollectionTyped(navigation).Contains(value.Entity);
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        public override bool AddToCollection(INavigation navigation, InternalEntityEntry value, bool forMaterialization)
-        {
-            if (navigation.TargetEntityType.ClrType == null)
-            {
-                return false;
-            }
-
-            var collection = GetOrCreateCollectionTyped(navigation);
-            if (!collection.Contains(value.Entity))
-            {
-                collection.Add(value.Entity);
-                return true;
-            }
-
-            return false;
-        }
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        public override bool RemoveFromCollection(INavigation navigation, InternalEntityEntry value)
-            => GetOrCreateCollectionTyped(navigation).Remove(value.Entity);
-    }
-}
