+++ src/EFCore/ChangeTracking/Internal/OriginalPropertyValues.cs
@@ -41,25 +41,6 @@ namespace Microsoft.EntityFrameworkCore.ChangeTracking.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public override bool TryGetValue<TValue>(string propertyName, out TValue value)
-        {
-            var property = InternalEntry.EntityType.FindProperty(propertyName);
-            if (property != null)
-            {
-                value = this.GetValue<TValue>(propertyName);
-                return true;
-            }
-
-            value = default(TValue);
-            return false;
-        }
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
         public override TValue GetValue<TValue>(IProperty property)
             => InternalEntry.GetOriginalValue<TValue>(InternalEntry.EntityType.CheckPropertyBelongsToType(property));
